storage/requirements.txt
connexion==2.14.2
apscheduler==3.10.4
requests==2.31.0
swagger-ui-bundle==0.0.9
SQLAlchemy==2.0.21
PyMySQL==1.1.0
pykafka==2.8.0